@@ -30,10 +30,10 @@
     "aphiria/console": "1.0.*@dev",
     "ext-mbstring": "*",
     "opulence/environments": "1.1.*@dev",
-    "php": "^7.3"
+    "php": "^7.4"
   },
   "require-dev": {
-    "friendsofphp/php-cs-fixer": "~2.14",
+    "friendsofphp/php-cs-fixer": "^3.0",
     "phpunit/phpunit": "~8.0"
   },
   "scripts": {
